Extract AdminClientToggle component for cleaner code organization
- Created new AdminClientToggle component for admin-only dropdown
- Moved all dropdown state and logic to the component
- Component handles click-outside behavior internally
- Dashboard page is now cleaner with 49 fewer lines
- Better separation of concerns and reusability
