+++ app/dashboard/page.tsx
@@ -6,6 +6,7 @@ import { useRouter } from 'next/navigation'
 import { User, Organization, Client, ClientOrganizationHistory } from '@/lib/supabase/types'
 import { SortField, SortDirection } from '@/lib/types/dashboard'
 import { formatCurrency, formatDate, formatFieldValue } from '@/app/utils/formatters'
+import { AdminClientToggle } from '@/app/components/dashboard/AdminClientToggle'
 
 
 
@@ -23,7 +24,6 @@ export default function DashboardPage() {
   const [sortDirection, setSortDirection] = useState<SortDirection>('asc')
   const [expandedRows, setExpandedRows] = useState<Set<string>>(new Set())
   const [orgDetails, setOrgDetails] = useState<Record<string, (ClientOrganizationHistory & { positions?: any[] }) | null>>({})
-  const [isDropdownOpen, setIsDropdownOpen] = useState(false)
   const router = useRouter()
   const supabase = createClientComponentClient()
 
@@ -144,24 +144,6 @@ export default function DashboardPage() {
     await supabase.auth.signOut()
     router.push('/')
   }
-
-  // Handle clicking outside dropdown to close it
-  useEffect(() => {
-    const handleClickOutside = (event: MouseEvent) => {
-      const target = event.target as HTMLElement
-      if (!target.closest('#client-dropdown-container')) {
-        setIsDropdownOpen(false)
-      }
-    }
-
-    if (isDropdownOpen) {
-      document.addEventListener('mousedown', handleClickOutside)
-    }
-
-    return () => {
-      document.removeEventListener('mousedown', handleClickOutside)
-    }
-  }, [isDropdownOpen])
 
   // Sorting logic
   const sortedOrganizations = useMemo(() => {
@@ -284,46 +266,15 @@ export default function DashboardPage() {
                 </h1>
               )}
               {userData?.client_uuid === '36fee78e-9bac-4443-9339-6f53003d3250' && (
-                <div id="client-dropdown-container" className="relative">
-                  <button
-                    onClick={() => setIsDropdownOpen(!isDropdownOpen)}
-                    className="p-2 text-gray-600 hover:text-gray-900 hover:bg-gray-100 rounded-md transition-colors"
-                    aria-label="Switch company"
-                  >
-                    <svg className="h-5 w-5" fill="none" viewBox="0 0 24 24" stroke="currentColor">
-                      <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M19 9l-7 7-7-7" />
-                    </svg>
-                  </button>
-                  
-                  {isDropdownOpen && (
-                    <div className="absolute left-0 mt-2 w-56 rounded-md shadow-lg bg-white ring-1 ring-black ring-opacity-5 z-50">
-                      <div className="py-1" role="menu" aria-orientation="vertical">
-                        <div className="px-4 py-2 text-xs text-gray-500 font-semibold uppercase tracking-wider">
-                          Switch Company
-                        </div>
-                        {clients.map((client) => (
-                          <button
-                            key={client.uuid}
-                            onClick={() => {
-                              setSelectedClientUuid(client.uuid)
-                              setSelectedClient(client)
-                              setIsDropdownOpen(false)
-                            }}
-                            className={`w-full text-left px-4 py-2 text-sm hover:bg-gray-100 ${
-                              selectedClientUuid === client.uuid ? 'bg-gray-50 font-medium' : ''
-                            }`}
-                            role="menuitem"
-                          >
-                            {client.name}
-                            {selectedClientUuid === client.uuid && (
-                              <span className="ml-2 text-primary-600">✓</span>
-                            )}
-                          </button>
-                        ))}
-                      </div>
-                    </div>
-                  )}
-                </div>
+                <AdminClientToggle
+                  clients={clients}
+                  selectedClientUuid={selectedClientUuid}
+                  selectedClient={selectedClient}
+                  onClientChange={(client) => {
+                    setSelectedClientUuid(client.uuid)
+                    setSelectedClient(client)
+                  }}
+                />
               )}
             </div>
             <div className="flex items-center space-x-4">
